[ uhc ] Add more primitives
