+++ src/full/Agda/Compiler/UHC/AuxASTUtil.hs
@@ -1,12 +1,18 @@
+{-# LANGUAGE CPP #-}
+-- has it's own module to break dependency loop
 module Agda.Compiler.UHC.AuxASTUtil
   ( bindExpr
+  , ctagCtorName
   )
 where
 
+import UHC.Light.Compiler.Core.API
 
-import Agda.Compiler.UHC.CompileState
 import Agda.Compiler.UHC.AuxAST
 import Agda.Compiler.UHC.Naming
+
+#include "undefined.h"
+import Agda.Utils.Impossible
 
 -- | Bind an expression to a fresh variable name
 bindExpr :: Monad m => Expr -> (HsName -> FreshNameT m Expr) -> FreshNameT m Expr
@@ -17,4 +23,6 @@ bindExpr expr f =
         v <- freshLocalName
         body <- f v
         return $ lett v expr body
-                
+
+ctagCtorName :: CTag -> HsName
+ctagCtorName = destructCTag __IMPOSSIBLE__ (\_ x _ _ -> x)
