+++ src/full/Agda/Compiler/UHC/Pragmas/Parse.hs
@@ -20,7 +20,7 @@ import qualified Data.Map as M
 import Agda.TypeChecking.Monad
 import Agda.Compiler.UHC.AuxAST
 import Agda.Compiler.UHC.Pragmas.Base
-import Agda.Compiler.UHC.Primitives
+import Agda.Compiler.UHC.Builtins
 
 import UHC.Light.Compiler.Core.API as CA
